Split the rendering from the webgl backend
