--- src/map/shaders3d/fill-fragment-shader.glsl
@@ -1,10 +1,0 @@
-
-#ifdef GL_ES
-    precision highp float;
-#endif
-    
-uniform vec3 uFillColor;
-
-void main() {
-    gl_FragColor = vec4(uFillColor, 1.0);
-}
