--- src/map/shaders3d/fill-vertex-shader.glsl
@@ -1,8 +1,0 @@
-
-attribute vec3 aVertexPosition;
-
-uniform mat4 uTransform;
-
-void main() {
-    gl_Position = uTransform * vec4(aVertexPosition, 1.0);
-}
